+++ src/program/util/actor.hpp
@@ -8,8 +8,8 @@ using ActorBase = engine::actor::ActorBase;
 namespace lotuskit::util::actor {
     void createSimple(const std::string &actorName);
     void createSimpleXYZ(const std::string &actorName, float x, float y, float z);
-    void createAndWatch(u8 slotIndex, const std::string &actorName);
-    void createAndWatchXYZ(u8 slotIndex, const std::string &actorName, float x, float y, float z);
+    void createAndWatch(size_t slotIndex, const std::string &actorName);
+    void createAndWatchXYZ(size_t slotIndex, const std::string &actorName, float x, float y, float z);
 
     // FIXME too many args, but AS arrays aren't wired to sead containers yet
     void setPosXYZ(ActorBase* actor, float x, float y, float z);
